没有类型推断的Java8 流下了时代的眼泪

+++ handler/src/test/java/io/netty/handler/stream/ChunkedWriteHandlerTest.java
@@ -225,7 +225,7 @@ public class ChunkedWriteHandlerTest {
         ch.writeAndFlush(input).syncUninterruptibly();
         assertTrue(ch.finish());
 
-        assertEquals(0, ch.readOutbound());
+        assertEquals(0L, (long)ch.readOutbound());
         assertNull(ch.readOutbound());
     }
